fix(lint): split force-link anchors + IM-self FFI decls under 2000-line cap
The pause_self/resume_self additions tipped perry-ext-http/src/lib.rs (2004)
and runtime_decls/stdlib_ffi.rs (2002) over the 2000-line CI cap. Move the
two new self-return decls into stdlib_ffi_part2.rs and extract the
force-link anchor table (mod + #[used] static) into a new force_link.rs
sibling. No behavior change — declarations and the linker anchor are
identical, just relocated.
