+++ crates/perry-codegen/src/runtime_decls/stdlib_ffi_part2.rs
@@ -21,6 +21,9 @@ pub(crate) fn declare_stdlib_ffi_part2(module: &mut LlModule) {
     );
     module.declare_function("js_node_http_im_http_version_major", DOUBLE, &[I64]);
     module.declare_function("js_node_http_im_http_version_minor", DOUBLE, &[I64]);
+    // #4975 — self-returning IncomingMessage pause()/resume() (return `this`).
+    module.declare_function("js_node_http_im_pause_self", I64, &[I64]);
+    module.declare_function("js_node_http_im_resume_self", I64, &[I64]);
 
     // ========== Commander CLI (#5137) ==========
     // `program.args` getter + `.argument(spec)` — kept here (not stdlib_ffi.rs)
